Adding simple python app package which forwards zmq messages
Python application indefinitely reads messages over an inbound zmq and
writes those messages plus a suffix to an outbound queue.

Pakcage includes both linting and unit tests. There are only enough unit
tests to give an example of general structure and how tests are run. The
intention is to show an example and not to provide full coverage.