Packaging simple app into docker image w/ consul template
Docker image uses supervisord to manage two long running processes:
  - simple_app
  - consul-template

Simple app is just the zmq forwarder while consul-template watches
for changes in KV pairs of an external consul service and updates
configuration files for simple_app. If consul-template renders a
new configuration file it will restart simple_app via supervisord
so that a new app will be started with the new configuration.

Along with these two processes is an event listener for supervisord.
The listener looks for whether any processes have exited or crashed
and if so sends a SIGTERM to supervisord. This causes the supervisord
and subsequently the container to exit.