Simple hello world version of a component test passes
The component tests setup a simple_app container, a consul server,
and a test harness container. The test harness container performs
read and writes to simple_app zmq sockets and verifies that changes
to the consul configuration keys are reflected in simple_app.

There is at least one arbitrary sleep in the component test which
will need to be addressed. This is to prevent the component test
from reading and writing the next values before consul-template
has had a chance to update the simple_app service.